Added functional wrapper methods to Configs
Added 'edit' methods for each of the configs which take a Consumer
of that type of config, and write out the config afterward.

Added 'getFrom' methods for each of the configs which take a
Function which takes that type of config as an argument and
returns some type R.

The hope is that by using these methods instead of the simple
'get' and 'write' methods, we will no longer forget to write
out configs after modifying them.