Change how to work with channel configs
Similar to the 'get' methods in Configs, Channel::getConfig may
lead to not writing out changes to the config, so it is being
deprecated and replaced by Channel::getFromConfig and
Channel::editConfig.

Added methods to get and edit channel configs which take a Channel
rather than a String representing the channel name; these methods
do not throw ChannelNotFoundException.