Add wrappers for configs
WrappedConfig wraps around a config object and has methods to edit
the enclosed config or retrieve data from it. The WrappedConfig
handles edits in a single thread so that concurrent calls to
WrappedConfig::edit do not produce inconsistent versions of the
config, one of which will overwrite the other. WrappedConfig
is designed with the knowledge that it is extended by
UpdatingConfig (thus WrappedConfig.config is not final, even
though it is never changed after creation in WrappedConfig).

UpdatingConfig adds the functionality to monitor the filesystem
for changes to the file associated with the config. It will
automatically update the config in memory if the config on disk is
modified in a valid way. Updates from disk occur in the same
thread as edits to help ensure consistency. If the config on disk
is removed, it will automatically write out the one in memory to
replace it. UpdatingConfig::startMonitoring must be called to
actually monitor the file for changes; it does not do so
automatically upon creation. UpdatingConfig::stopMonitoring can
be called to stop monitoring the filesystem for changes to the
file.