Add WrappedConfig.getExactly()
Added method to WrappedConfig to get a value from the update
thread, to guarantee that it is the most up-to-date value. Added
methods to Configs to get the exact Service, either with possible
Exception throws or as an Optional. Used new Configs method to
get Service when starting the bot.