Clean up Channel
Added CooldownManager and ChannelScheduleManager classes. Copied
some methods from Channel to those classes, and deprecated those
methods in Channel. Methods moved to ChannelScheduleManager were
given package-private access, to prevent them from being called
from outside of the Schedules API.

The goal of this commit is to remove methods from Channel in
order to reduce clutter, and to reduce the number of methods to be
forwarded in the future ChannelProxy class.