Rewrite channel APIs
The Channels class/API provided convenient methods for joining and
leaving channels, but had major encapsulation problems. It added
Channels to and removed them from a map which was publicly visible
in the Bot. Nothing forced clients to call the methods from
Channels, and thus a channel could end up in an incosistent state
(joined for the Channel object, but not connected to the service).
Additionally, the Channels API exposed the Channel object, which
allowed clients to call Channel::leave without disconnecting from
the service, leaving the channel in the reverse of the previous
inconsistent state.

The Bot interface now has a channelManager() method which returns
a ChannelManager object. ChannelManager is virtually identical to
the Channels class, except that all of its methods are instance
methods linked to the now hidden Map of Channels in Bot, and that
its getting methods return ChannelProxy instead of Channel.
ChannelProxy is a wrapper class around a Channel, which exposes
all of its methods, except for join() and leave().

The combination of these changes means that attempts to join or
leave a channel are forced to be done through the ChannelManager
API, and the channel will be in a consistent join state. The
Channels class is now deprecated.