Clean up config APIs
Extracted interface from WrappedConfig, and added method to
UpdatingConfig to return a view of it as a WrappedConfig. This
protects methods like startMonitoring() in UpdatingConfig, but
exposes get and edit methods.

Deprecated getFrom and edit methods in Configs, and replaced them
with getters for WrappedConfigs of all of the configs. Did the same
for get and edit methods in Channel and ChannelProxy.