Add basics of event framework
Change event handling to use a Guava EventBus. Add Bot::eventBus,
and change message handlers for Beam and Twitch to use that instead
of Bot::invokeMessageHandlers.

Add a bunch of events which are not currently in use in preparation
for future event framework. Add AbstractServiceChannel and
ChannelState to prepare for monitoring whether a channel is live
and firing events based on that as well.