Fix concurrency problem changing Service
Fixed problems where Configs::getAccount would return an Account
with the wrong data due to either the private 'account' field not
being propogated to other threads, or because the updated Service
in the 'generalConfig' field was not being propogated to other
threads.

Added WrappedConfig::updateAndAwait, and refactored ::update to
::updateLater. This was done to fix a problem in Control where,
when reloading the GeneralConfig, it would queue an update and then
update the Account based on the GeneralConfig when the latter might
not have actually been updated yet.