Improve PluginData initialization
Improve SafeInitializer so that it prevents plugins which do not
have an initialization loop from initializing the same data
multiple times.

Add method to Data (and change its constructor signature) so it can
generate a PluginDataMap from a set of ContentPlugins and a
function to get a DataFactory from each plugin.